requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.32.3
selenium-stealth==1.0.6 